Use a separate thread to copy scan data to application buffer and fix broken PNG writer
- Separate thread prevents the scanner from stalling.
- Code was deleted by mistake from the PNG writer.
- Make sure all writers cleans themselves properly after writing is completed
- Make sure very large images (>2^31 bytes) are handled properly

+++ src/PreviewScanProcess.hpp
@@ -41,13 +41,13 @@ namespace Gorfector
             return true;
         }
 
-        void GetBuffer(SANE_Byte *&outBuffer, int &outMaxReadLength) override
+        void GetBuffer(SANE_Byte *&outBuffer, size_t &outMaxReadLength) override
         {
             auto previewPanelUpdater = PreviewState::Updater(m_PreviewState);
             previewPanelUpdater.GetReadBuffer(outBuffer, outMaxReadLength);
         }
 
-        void CommitBuffer(int32_t readLength) override
+        void CommitBuffer(size_t readLength) override
         {
             auto previewPanelUpdater = PreviewState::Updater(m_PreviewState);
             previewPanelUpdater.CommitReadBuffer(readLength);
